feature(berita): handling not found berita and berita image

+++ berita/src/main/java/com/myjisc/berita/restcontroller/BeritaRestController.java
@@ -43,25 +43,31 @@ public class BeritaRestController {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(responseBody);
         }
 
-        if (file != null) {
-            var beritaFromDTO = beritaMapper.createRestBeritaDTOToBerita(beritaRequestDTO);
-            var berita = beritaRestService.createRestBerita(beritaFromDTO, file);
-
+        try {
+            if (file != null) {
+                var beritaFromDTO = beritaMapper.createRestBeritaDTOToBerita(beritaRequestDTO);
+                var berita = beritaRestService.createRestBerita(beritaFromDTO, file);
+    
+                Map<String, Object> responseBody = new HashMap<>();
+                responseBody.put("status", "success");
+    
+                responseBody.put("data", berita);
+    
+                return ResponseEntity.status(HttpStatus.OK).body(responseBody);
+            } else {
+                var beritaFromDTO = beritaMapper.createRestBeritaDTOToBerita(beritaRequestDTO);
+                var berita = beritaRestService.createRestBerita(beritaFromDTO);
+    
+                Map<String, Object> responseBody = new HashMap<>();
+                responseBody.put("status", "success");
+                responseBody.put("data", berita);
+    
+                return ResponseEntity.status(HttpStatus.OK).body(responseBody);
+            }
+        } catch (Exception e) {
             Map<String, Object> responseBody = new HashMap<>();
-            responseBody.put("status", "success");
-
-            responseBody.put("data", berita);
-
-            return ResponseEntity.status(HttpStatus.OK).body(responseBody);
-        } else {
-            var beritaFromDTO = beritaMapper.createRestBeritaDTOToBerita(beritaRequestDTO);
-            var berita = beritaRestService.createRestBerita(beritaFromDTO);
-
-            Map<String, Object> responseBody = new HashMap<>();
-            responseBody.put("status", "success");
-            responseBody.put("data", berita);
-
-            return ResponseEntity.status(HttpStatus.OK).body(responseBody);
+            responseBody.put("message", "Check your input again");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseBody);
         }
     }
 
@@ -69,6 +75,13 @@ public class BeritaRestController {
     public ResponseEntity viewAllBerita() {
         try {
             List<Berita> listAvailableBerita = beritaRestService.retrieveRestAvailableBerita();
+
+            if (listAvailableBerita.isEmpty()) {
+                Map<String, Object> responseBody = new HashMap<>();
+                responseBody.put("message", "Data not found");
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseBody);
+            }
+
             Map<String, Object> responseBody = new HashMap<>();
             responseBody.put("status", "success");
 
@@ -90,6 +103,12 @@ public class BeritaRestController {
         try {
             var berita = beritaRestService.getRestBeritaById(idBerita);
 
+            if (berita == null) {
+                Map<String, Object> responseBody = new HashMap<>();
+                responseBody.put("message", "Berita not found");
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseBody);
+            }
+
             Map<String, Object> responseBody = new HashMap<>();
             responseBody.put("status", "success");
 
@@ -108,38 +127,54 @@ public class BeritaRestController {
             @Valid @RequestBody @ModelAttribute UpdateBeritaRequestDTO beritaRequestDTO,
             @RequestPart(value = "image", required = false) MultipartFile file, BindingResult bindingResult)
             throws IOException {
+        
         if (bindingResult.hasErrors()) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid data");
         }
 
-        if (file != null) {
-            var beritaFromDTO = beritaMapper.updateRestBeritaDTOToBerita(beritaRequestDTO);
-            beritaFromDTO.setIdBerita(id);
-            var berita = beritaRestService.updateRestBerita(beritaFromDTO, file);
-
+        try {
+            if (file != null) {
+                var beritaFromDTO = beritaMapper.updateRestBeritaDTOToBerita(beritaRequestDTO);
+                beritaFromDTO.setIdBerita(id);
+                var berita = beritaRestService.updateRestBerita(beritaFromDTO, file);
+    
+                Map<String, Object> responseBody = new HashMap<>();
+                responseBody.put("status", "success");
+    
+                responseBody.put("data", berita);
+    
+                return ResponseEntity.status(HttpStatus.OK).body(responseBody);
+            } else {
+                var beritaFromDTO = beritaMapper.updateRestBeritaDTOToBerita(beritaRequestDTO);
+                beritaFromDTO.setIdBerita(id);
+                var berita = beritaRestService.updateRestBerita(beritaFromDTO);
+    
+                Map<String, Object> responseBody = new HashMap<>();
+                responseBody.put("status", "success");
+                responseBody.put("data", berita);
+    
+                return ResponseEntity.status(HttpStatus.OK).body(responseBody);
+            }
+        } catch (Exception e) {
             Map<String, Object> responseBody = new HashMap<>();
-            responseBody.put("status", "success");
-
-            responseBody.put("data", berita);
-
-            return ResponseEntity.status(HttpStatus.OK).body(responseBody);
-        } else {
-            var beritaFromDTO = beritaMapper.updateRestBeritaDTOToBerita(beritaRequestDTO);
-            beritaFromDTO.setIdBerita(id);
-            var berita = beritaRestService.updateRestBerita(beritaFromDTO);
-
-            Map<String, Object> responseBody = new HashMap<>();
-            responseBody.put("status", "success");
-            responseBody.put("data", berita);
-
-            return ResponseEntity.status(HttpStatus.OK).body(responseBody);
+            responseBody.put("message", "Check your input again");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseBody);
         }
+
+       
     }
 
     @GetMapping(value = "/{id}/image")
     public ResponseEntity getBeritaImage(@PathVariable("id") Long id) {
         try {
             byte[] image = beritaRestService.getImageBerita(id);
+            
+            if (image == null) {
+                Map<String, Object> responseBody = new HashMap<>();
+                responseBody.put("message", "Image not found");
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseBody);
+            }
+
             return ResponseEntity.status(HttpStatus.OK).contentType(MediaType.IMAGE_PNG).body(image);
         } catch (Exception e) {
             Map<String, Object> responseBody = new HashMap<>();
@@ -152,6 +187,13 @@ public class BeritaRestController {
     public ResponseEntity deleteBerita(@PathVariable("id") Long id) {
         try {
             var berita = beritaRestService.getRestBeritaById(id);
+
+            if (berita == null) {
+                Map<String, Object> responseBody = new HashMap<>();
+                responseBody.put("message", "Artikel not found");
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseBody);
+            }
+
             beritaRestService.deleteRestBerita(berita);
 
             Map<String, Object> responseBody = new HashMap<>();
